requirements.txt
annotated-types==0.6.0
anyio==4.2.0
click==8.1.7
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.108.0
h11==0.14.0
idna==3.6
motor==3.3.2
pyasn1==0.5.1
pycryptodome==3.19.1
pydantic==2.5.3
python-dotenv==1.0.0
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.32.0.post1
typing_extensions==4.9.0
uvicorn==0.25.0
pymongo
